+++ src/index.ts
@@ -1,15 +1,14 @@
 import readline = require("readline");
-import { Command, Commands, Action, StrictCommands } from "./definitions";
+import { Command, Commands, Action, Parameter } from "./definitions";
 import { parseOptions } from "./option-parser";
 import { printCommandHelp, printOverviewHelp } from "./help-gen";
-import { parseCommand } from "./command-parser";
+import { findPromptedCommand } from "./command-parser";
 import { parseParameters } from "./parameter-parser";
-import { commandToStrictCommand } from "./helpers";
 
 export * from "./definitions";
 
 export class CLI {
-    private readonly cmdRegistry: StrictCommands = {};
+    private readonly cmds: Commands = {};
     private readonly readline: readline.ReadLine;
     private delimiter = "$> ";
     private isActive = false;
@@ -27,24 +26,50 @@ export class CLI {
         this.readline.pause();
     }
 
-    private async executeCommand(commandStr: string) {
-        const pieces = commandStr.split(" ");
-        if (pieces[0] === "help") return this.help(pieces.slice(1));
-        const command = parseCommand(pieces, this.cmdRegistry);
-        if (!command) return this.invalidCommand();
-        const options = parseOptions(command.command, command.remainingPieces);
-        if (!options) return this.help(pieces);
-        const params = parseParameters(command.command, options.remainingPieces);
-        if (!params) return this.help(pieces);
-        return new Promise(resolve => {
-            const prom = command.command.action(params, options.options, resolve);
-            if (prom instanceof Promise) {
-                prom.then(resolve).catch(e => {
-                    console.log(e);
-                    resolve();
-                });
+    private paramIsRequired(param: Parameter | string) {
+        if (typeof param === "string") return true;
+        return !param.optional && !param.rest;
+    }
+
+    private isRestParameter(param: Parameter | string) {
+        return !(typeof param === "string") && param.rest;
+    }
+
+    private checkCommandForErrors(command: Command | Action) {
+        if (command instanceof Function) return;
+        if (!command.parameters) return;
+
+        let hasHadOptional = false;
+        command.parameters.forEach((param, index, parameters) => {
+            const isRequired = this.paramIsRequired(param);
+
+            if (isRequired && hasHadOptional) {
+                throw new Error(`Invalid parameter order. Required parameter after optional.`);
+            }
+
+            if (!isRequired) hasHadOptional = true;
+
+            if (this.isRestParameter(param) && (index + 1) !== parameters.length) {
+                throw new Error(`Rest parameter be last`);
             }
         });
+    }
+
+    private async executeCommand(input: string): Promise<void> {
+        const pieces = input.split(" ");
+
+        if (pieces[0] === "help") return this.help(pieces.slice(1));
+
+        const parsedCmd = findPromptedCommand(pieces, this.cmds);
+        if (!parsedCmd) return this.invalidCommand();
+
+        const options = parseOptions(parsedCmd.command, parsedCmd.remainingPieces);
+        if (!options) return this.help(pieces);
+
+        const params = parseParameters(parsedCmd.command, options.remainingPieces);
+        if (!params) return this.help(pieces);
+
+        return parsedCmd.command.action(params, options.options);
     }
 
     private invalidCommand() {
@@ -58,24 +83,22 @@ export class CLI {
         });
     }
 
-    private startREPL() {
-        this.prompt()
-        .then(commandStr => this.executeCommand(commandStr))
-        .catch(e => console.log(e))
-        .then(() => {
-            if (this.isActive) this.startREPL();
-        });
+    private async startREPL() {
+        while (this.isActive) {
+            const input = await this.prompt();
+            await this.executeCommand(input).catch(e => console.log(e));
+        }
     }
 
     private help(commandPieces: string[]): void {
         if (commandPieces.length === 0) {
             printOverviewHelp({
-                info: this.info, name: this.name, version: this.version, commands: this.cmdRegistry
+                info: this.info, name: this.name, version: this.version, commands: this.cmds
             });
             return;
         }
 
-        const commandOpts = parseCommand(commandPieces, this.cmdRegistry);
+        const commandOpts = findPromptedCommand(commandPieces, this.cmds);
 
         if (!commandOpts) return this.help([]);
 
@@ -86,56 +109,95 @@ export class CLI {
     }
 
     /** Set the cli delimiter */
-    setDelimiter(delimiter: string) {
+    setDelimiter(delimiter: string): this {
         this.delimiter = delimiter;
+        return this;
     }
 
     /** Register a command */
-    command(command: string, opts: Command): void;
-    command(command: string, action: Action): void;
-    command(command: string, opts: Command | Action): void;
-    command(command: string, opts: Command | Action): void {
-        this.cmdRegistry[command] = commandToStrictCommand(opts);
+    addCommand(name: string, command: Command): this;
+    addCommand(name: string, action: Action): this;
+    addCommand(name: string, command: Command | Action): this;
+    addCommand(name: string, command: Command | Action): this {
+        this.checkCommandForErrors(command);
+        this.cmds[name] = command;
+        return this;
     }
 
     /** Register multiple commands at once (Alias for registerCommands) */
-    commands(commands: Commands) {
-        this.registerCommands(commands);
+    addCommands(commands: Commands): this {
+        for (const command in commands) {
+            this.addCommand(command, commands[command]);
+        }
+        return this;
     }
 
-    /** Register multiple commands at once */
-    registerCommands(commands: Commands) {
-        for (const command in commands) {
-            this.command(command, commands[command]);
-        }
+    removeCommand(command: string): this {
+        delete this.cmds[command];
+        return this;
     }
 
     /** Show the CLI */
-    async show() {
+    show(): this {
         this.readline.resume();
         this.isActive = true;
         this.startREPL();
+        return this;
     }
 
     /** Hide the cli */
-    hide() {
+    hide(): this {
         this.readline.pause();
         this.isActive = false;
+        return this;
     }
 
-    setVersion(val: string) {
+    setVersion(val: string): this {
         this.version = val;
+        return this;
     }
 
-    setInfo(val: string) {
+    setInfo(val: string): this {
         this.info = val;
+        return this;
     }
 
-    setName(val: string) {
+    setName(val: string): this {
         this.name = val;
+        return this;
     }
 
     hasCommand(cmd: string): boolean {
-        return !!this.cmdRegistry[cmd];
+        return !!this.cmds[cmd];
+    }
+
+    /**
+     * Register a command.
+     *
+     * @deprecated - Use addCommand instead.
+     */
+    command(name: string, opts: Command | Action): this {
+        console.warn("command is deprecated - use addCommand instead");
+        return this.addCommand(name, opts);
+    }
+
+    /**
+     * Register multiple commands at once (Alias for registerCommands).
+     *
+     * @deprecated - Use addCommands instead.
+     **/
+    commands(name: Commands): this {
+        console.warn("commands is deprecated - use addCommands instead");
+        return this.addCommands(name);
+    }
+
+    /**
+     * Register multiple commands at once (Alias for registerCommands).
+     *
+     * @deprecated - Use addCommands instead.
+     **/
+    registerCommands(name: Commands): this {
+        console.warn("registerCommands is deprecated - use addCommands instead");
+        return this.registerCommands(name);
     }
 }